api/requirements.txt
fastapi==0.136.0
uvicorn==0.46.0
mysql-connector-python==9.6.0
motor==3.7.1
